Rename "deciles" app to something a little clearer
...as there are also deciles on the measures output
