+++ index.py
@@ -20,7 +20,7 @@ def setup_app_and_layout():
 def setup_callbacks():
     import apps.base
 
-    import apps.deciles
+    import apps.analyse
     import apps.heatmap
     import apps.datatable
     import apps.measure
